Update histories tests
